Convert rdfaparser.js to require()
